pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent"
version = "0.4.0"
description = "Multi-agent orchestration framework for IDE-based AI coding assistants"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4,<1",
    "langgraph-checkpoint-sqlite>=2,<3",
    "pydantic>=2.5,<3",
    "pyyaml>=6,<7",
    "jinja2>=3.1,<4",
    "click>=8,<9",
    "watchdog>=4,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]
mcp = [
    "mcp>=1,<2",
]

[project.scripts]
ma = "multi_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
